fix: command/circuitCommand/exportCommand fix ExportJSON class with new circuit utilities

+++ quantum-core/src/main/java/control/command/circuitCommand/exportCommand/ExportJSON.java
@@ -11,25 +11,17 @@ import java.util.Map;
 
 public class ExportJSON implements QuantumCommand {
     private final QuantumCircuit circuit;
-    public QuantumCircuitCLIDisplay utils = new QuantumCircuitCLIDisplay();
 
-    public ExportJSON(QuantumCircuit circuit) {
-        this.circuit = circuit;
-    }
+    public ExportJSON(QuantumCircuit circuit) {this.circuit = circuit;}
 
     @Override
     public void execute() {
         ObjectMapper mapper = new ObjectMapper();
         try {
             Map<String, Object> circuitData = new HashMap<>();
-            Map<String, String> wire = new HashMap<>();
-            for (int i=0;i<circuit.getCircuit().size();i++) {
-                String nWire = "wire-" + i;
-                wire.put(nWire, utils.formatWire(circuit.getCircuit().get(i)));
-                circuitData.put("circuit", wire);
-            }
-            circuitData.put("numOfQubits", circuit.getnQubits());
-            circuitData.put("depth", circuit.getDepth());
+            circuitData.put("circuit", circuit.toString());
+            circuitData.put("nQubit", circuit.getNQubits());
+            circuitData.put("nLayer", circuit.getDepth());
             mapper.writeValue(new File("./exported_circuit.json"), circuitData);
         } catch (IOException e) {
             throw new RuntimeException("Error creating json response: "+ e.getMessage(), e);
@@ -39,18 +31,12 @@ public class ExportJSON implements QuantumCommand {
     @Override
     public void undo() {
         File exportedFile = new File("./exported_circuit.json");
-        if (exportedFile.exists()) {
-            exportedFile.delete();
-        }
+        if (exportedFile.exists()) exportedFile.delete();
     }
 
     @Override
-    public boolean canUndo() {
-        return true;
-    }
+    public boolean canUndo() {return true;}
 
     @Override
-    public void redo() {
-        execute();
-    }
+    public void redo() {execute();}
 }
